+++ src/loop.h
@@ -53,8 +53,8 @@ extern LOOP_MISSION_STATE		loopMissionState;
 // this is set by scrStartMission to say what type of new level is to be started
 extern LEVEL_TYPE nextMissionType;
 
-extern unsigned int loopPieCount;
-extern unsigned int loopPolyCount;
+extern size_t loopPieCount;
+extern size_t loopPolyCount;
 
 GAMECODE gameLoop();
 void videoLoop();
